backend/pyproject.toml
[project]
name = "asr-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "dashscope>=1.25.14",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]
